Revert "ADM-519 feat: finishing the design of error page"
This reverts commit fdcfa86a3d665e1cb64d3b98b4db821b83f57bb0.
