+++ frontend/src/components/ErrorMessage/index.tsx
@@ -1,0 +1,3 @@
+export const ErrorMessage = () => {
+  return 'errorMessage'
+}
